backend/requirements.txt
# DealMotion Marketing Engine - Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# AI / LLM
anthropic>=0.18.0

# Text-to-Speech
httpx>=0.26.0

# YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Inngest (Workflow Orchestration)
inngest>=0.4.0

# Database (Supabase)
supabase>=2.3.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Utilities
python-multipart>=0.0.6

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0

